update: integrate wdmpnn rdkit funcs into ext_rdkit.py
